requirements.txt
textual>=0.47.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
